% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binS.R
\name{generate.binS}
\alias{generate.binS}
\title{Generates binary data for model Binary:M with lambda depending 
on a scalar covariate.}
\usage{
generate.binS(param, R, J, Tmax, covar)
}
\arguments{
\item{param}{A vector comprised of the Poisson rate lambda, and the detection hazard, h.}

\item{R}{The number of sites.}

\item{J}{The number of occasions (assumed the same for all sites).}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}

\item{covar}{A (scalar) covariate on which the hazard depends linearly (assumed to be 
the same for all sites).}
}
\value{
Returns an \code{R} by \code{J} matrix of binary values, with a 1 representing
detection and a 0 representing no detection.
}
\description{
Generates Poisson random variables for each of the \code{R} sites, corresponding to the 
number of animals in each site. Then generates Bernoulli random variables for each of 
\code{J} occasions that each site is surveyed, assuming a constant hazard of detection 
for each animal.
}
\examples{
###### NOT YET UPDATED #########
# setting
Rsites=100  #number of sites
Jsites=5    #multiple visits
Tsearch=3 #maximum time
#true parameters
lamt = 2
ht=0.4620981
paramt=c(lamt,ht)

bnm=as.matrix(generate.binM(paramt,R=Rsites,J=Jsites,Tmax=Tsearch))
str(bnm)
head(bnm)

}
