% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binS.R
\name{generate.binScov}
\alias{generate.binScov}
\title{Generates binary data for model Binary:S with lambda depending 
on a covariate.}
\usage{
generate.binScov(param, R, Tmax, covar)
}
\arguments{
\item{param}{A vector comprised of parameters \eqn{b_0}{b0}, \eqn{b_1}{b1}, 
and the log of the detection hazard \eqn{h}{h} (in that order), where the log of the Poisson rate 
lambda is equal to \eqn{b_0+b_1x}{b0+b1*x} and \eqn{x}{x} is the covariate \code{covar}.}

\item{R}{The number of sites.}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}

\item{covar}{A vector covariate of length \code{R} on which the expected number of 
animals in the site depends linearly (assumed to be the same for all occasions).}
}
\value{
Returns an \code{R} by 1 matrix of binary values, with a 1 representing
detection and a 0 representing no detection.
}
\description{
Generates Poisson random variables for each of the \code{R} sites, corresponding to the 
number of animals in each site and depending on the covariate value attached to the site. 
Then generates Bernoulli random variables for each site, assuming a constant hazard of 
detection for each animal.
}
\examples{
Rsites=100. #number of sites
Tsearch=3 #maximum time

#true parameters
b0t =0.660878  #true intercept for log(lambda) 
b1t =0.255413 #true slope for log(lambda)
ht=0.4620981  #rate parameter

# simulation of the covariate vegHt over R sites 
# hence the site-dependent abundance for the true parameters b0t and b1t
set.seed(123) # for reprodicibility
x=rnorm(Rsites,0,1)
lambda=exp(b0t+b1t*x)
mean(lambda)

# true probabilites
1-exp(-ht*Tsearch)      #P(individual detected)
1-exp(-mean(lambda)) #P(site occupancy)
paramt=c(b0t, b1t,ht)

# data for Binary:S
bns=as.matrix(generate.binScov(paramt,R=Rsites,Tmax=Tsearch,covar=x))
str(bns)
head(bns)

}
