% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binT1M.R
\name{generate.binT1M}
\alias{generate.binT1M}
\title{Generates time to first detection data for model BinaryT1:M with lambda constant}
\usage{
generate.binT1M(param, R, J, Tmax)
}
\arguments{
\item{param}{A vector comprised of the Poisson rate lambda, and the detection hazard, h.}

\item{R}{The number of sites.}

\item{J}{The number of occasions (assumed the same for all sites).}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}
}
\value{
Returns an \code{R} by \code{J} matrix of times to first detection values, with 
a time equal to \code{Tmax} indicating no detection.
}
\description{
Generates Poisson random variables for each of the \code{R} sites, corresponding to the 
number of animals in each site. Then generates an exponential random variable for 
each site surveyed, with hazard \eqn{hn}{hn},  where \eqn{h}{h} is the hazard of 
detection for a single animal.
}
\examples{
# setting
Rsites=100  #number of sites
Jsites=5    #multiple visits
Tsearch=3 #maximum time
#true parameters
lamt = 2
ht=0.4620981
paramt=c(lamt,ht)

t1m=as.matrix(generate.binT1M(paramt,R=Rsites,J=Jsites,Tmax=Tsearch))
str(t1m)
head(t1m)

}
