% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binT1S.R
\name{generate.binT1S}
\alias{generate.binT1S}
\title{Generates time to first detection data for model BinaryT1:S with constant lambda.}
\usage{
generate.binT1S(param, R, Tmax)
}
\arguments{
\item{param}{A vector comprised of the Poisson rate parameter lambda, and the detection hazard
\eqn{h}{h}.}

\item{R}{The number of sites.}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}
}
\value{
Returns an \code{R} by 1 matrix containing the times of
first detections in each site, with a time equal to \code{Tmax} indicating no detection.
}
\description{
Generates Poisson random variables \eqn{n}{n} for each of the \code{R} sites, corresponding to the 
number of animals in each site. 
Then generates an exponential random variable for each site surveyed, with hazard \eqn{hn}{hn}, 
where \eqn{h}{h} is the hazard of detection for a single animal.
}
\examples{
Rsites=100. #number of sites
Tsearch=3 #maximum time

#true parameters
lamt = 2
ht=0.4620981
paramt=c(lamt,ht)

# data for BinaryT1:S
bns=as.matrix(generate.binT1S(paramt,R=Rsites,Tmax=Tsearch))
str(bns)
head(bns)

}
