% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMT1.R
\name{generate.countMT1}
\alias{generate.countMT1}
\title{Generates count data for model CountT1:M with lambda constant}
\usage{
generate.countMT1(param, R, J, Tmax)
}
\arguments{
\item{param}{A vector comprised of the Poisson rate lambda, and the detection hazard, h.}

\item{R}{The number of sites.}

\item{J}{The number of occasions (assumed the same for all sites).}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}
}
\value{
Returns an \code{R} by \code{2J} matrix with first \code{J} columns being the numbers 
of detections in each site on each occasion and the remaining \code{J} columns being the 
shortest time to detection for all detected  animals with in the site on the occasion 
(with zero representing no detections).
}
\description{
Generates Poisson random variables for each of the \code{R} sites, corresponding to the 
number of animals in each site on each of \code{J} occasions. 
Then generates exponential random variables for each ainimal in each site, 
assuming a constant hazard of detection for each animal.
}
\examples{
# setting
Rsites=100  #number of sites
Jsites=5    #multiple visits
Tsearch=3 #maximum time
#true parameters
lamt = 2
ht=0.4620981
paramt=c(lamt,ht)

# data for CountT1:M
cntm=as.matrix(generate.countMT1(paramt,R=Rsites,J=Jsites,Tmax=Tsearch))
str(cntm)
head(cntm)

}
