% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PcountM.R
\name{generate.PcountM}
\alias{generate.PcountM}
\title{Generates count data for model PCount:M with lambda constant}
\usage{
generate.PcountM(param, R, J, Tmax)
}
\arguments{
\item{param}{A vector comprised of the Poisson rate lambda, and the detection hazard, h.}

\item{R}{The number of sites.}

\item{J}{The number of occasions.}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}
}
\value{
Returns an \code{R} by \code{J} matrix of counts.
}
\description{
Generates Poisson random variables for each of the \code{R} sites, corresponding to the 
number of animals in each site. Then generates Poisson random variables for each 
site, on each of \code{J} occassion, given the number of animals present, assuming a 
constant hazard of detection for each animal.
}
\examples{
# seed
set.seed(123)
# setting
Rsites=100. #number of sites
Tsearch=3 #maximum time
Joccs=5 #number of occasions
#true parameters
lamt=2
gt=0.4620981
paramt=c(lamt,gt)

ymat=generate.PcountM(paramt, R=Rsites, J=Joccs, Tmax=Tsearch)
str(ymat)
head(ymat)

}
