% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PcountM.R
\name{generate.PcountMcov}
\alias{generate.PcountMcov}
\title{Generates count data for model PCount:M with lambda depending on a covariate.}
\usage{
generate.PcountMcov(param, R, J, Tmax, covar)
}
\arguments{
\item{param}{A vector comprised of the Poisson rate lambda, and the detection hazard, h.}

\item{R}{The number of sites.}

\item{J}{The number of occasions}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}

\item{covar}{A vector covariate of length \code{R} on which the expected number of 
animals in the site depends linearly (assumed to be the same for all occasions).}
}
\value{
Returns an \code{R} by \code{J} matrix of counts.
}
\description{
Generates Poisson random variables for each of the \code{R} sites, corresponding to the 
number of animals in each site and depending on the covariate value attached to the site. 
Then generates Poisson random variables for each site, oneach of \code{J} occasions, 
given the number of animals present, assuming a constant hazard of detection for each animal.
}
\examples{
# seed
set.seed(321)
# setting
Rsites=100. #number of sites
Joccs=5 # number of occasions
Tsearch=3 #maximum time
#true parameters
b0t =0.660878  
b1t =0.255413
gt=0.4620981
# simulation of the covariate x over R sites for true parameters
x=rnorm(R,0,1)
hist(x)
lambda=exp(b0t+b1t*x)
mean(lambda)
# true parameters
paramt=c(b0t,b1t,gt)

# generate
ymat=generate.PcountMcov(paramt, R=Rsites, J=Joccs, Tmax=Tsearch, covar=x)
str(ymat)
head(ymat)

}
