% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Karoodata}
\alias{Karoodata}
\title{Data from survey in the Karoo}
\format{

}
\source{
https://doi.org/10.5061/dryad.msbcc2fv7 where it is licensed under a CC0 1.0 Universal (CC0 1.0) Public Domain Dedication license.
}
\usage{
Karoodata
}
\description{
Data from a bird survey in the Karoo region of South Africa.
}
\details{
These data were collected by Alan T. K. Lee and consist of 10 minute point counts during which all detected birds were recorded, along with the time since the start of the survey and distance between the bird and the observer. The surveys covered 61 sites - pentads. Each pentad was surveyed multiple times. The repeat surveys consisted of spatially replicated point counts. More detail on the data collection methods and study area can be found in Henry et al 2020 (reference below) and Lee et al 2018 (Ostrich 89: 363-372; DOI: 10.2989/00306525.2018.1553804)
}
\examples{
 data(Karoodata)
}
\references{
Henry, Dominic; Lee, Alan; Altwegg, Res (2020), Can time-to-detection models with fewer 
survey replicates provide a robust alternative to traditional site-occupancy models? 
Methods in Ecology and Evolution, 11: 643-655. 
https://besjournals.onlinelibrary.wiley.com/doi/10.1111/2041-210X.13379
}
