% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binM.R
\name{nll.binM}
\alias{nll.binM}
\title{Evaluates the negeative log-likelihood for model Binary:M with lambda constant}
\usage{
nll.binM(param, R, J, Tmax, dat)
}
\arguments{
\item{param}{A vector comprised of the log of the Poisson rate lambda, and the 
log of the detection hazard, h.}

\item{R}{The number of sites.}

\item{J}{The number of occasions (assumed the same for all sites).}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}

\item{dat}{An \code{R} by \code{J} matrix of binary values, with a 1 representing
detection and a 0 representing no detection}
}
\value{
Returns the negative log-likelihood function evaluated at the parameter values
passed in \code{param}.
}
\description{
Evaluates the negeative log-likelihood for model Binary:M, assuming constant lambda, 
given initial parameter estimates and binary data from a multiple-occasion survey.
}
\examples{
# setting
Rsites=100  #number of sites
Jsites=5    #multiple visits
Tsearch=3 #maximum time
# true parameters
lamt = 2
ht=0.4620981
paramt=c(lamt,ht)

# data for Binary:M
set.seed(123) # for reproducibility
bnm=as.matrix(generate.binM(paramt,R=Rsites,J=Jsites,Tmax=Tsearch))
# optimize
init.paramt=c(log(lamt),log(ht))
fit.binM=optim(init.paramt,nll.binM,R=Rsites,J=Jsites,Tmax=Tsearch,dat=bnm)
estpar.binM=fit.binM$par
# compare estimates and true parameters
exp(estpar.binM)
paramt 

}
