% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binT1M.R
\name{nll.binT1M}
\alias{nll.binT1M}
\title{Evaluates the negeative log-likelihood for model BinaryT1:M with lambda constant}
\usage{
nll.binT1M(param, R, J, Tmax, dat)
}
\arguments{
\item{param}{A vector comprised of the log of the Poisson rate lambda, and the 
log of the detection hazard, h.}

\item{R}{The number of sites.}

\item{J}{The number of occasions (assumed the same for all sites).}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}

\item{dat}{An \code{R} by \code{J} matrix of time to first detection values, with a 
time of \code{Tmax} representing no detection}
}
\value{
Returns the negative log-likelihood function evaluated at the parameter values
passed in \code{param}.
}
\description{
Evaluates the negeative log-likelihood for model BinaryT1:M, assuming constant lambda, 
given initial parameter estimates and time to first detection data from a 
multiple-occasion survey.
}
\examples{
# setting
Rsites=100  #number of sites
Jsites=5    #multiple visits
Tsearch=3 #maximum time
# true parameters
lamt = 2
ht=0.4620981
paramt=c(lamt,ht)

# data for Binary:M
set.seed(123) # for reproducibility
t1m=as.matrix(generate.binT1M(paramt,R=Rsites,J=Jsites,Tmax=Tsearch))
# optimize
init.paramt=c(log(lamt),log(ht))
fit.binT1M=optim(init.paramt,nll.binT1M,R=Rsites,J=Jsites,Tmax=Tsearch,dat=t1m)
estpar.binT1M=fit.binT1M$par
# compare estimates and true parameters
exp(estpar.binT1M)
paramt 

}
