% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binT1S.R
\name{nll.binT1S}
\alias{nll.binT1S}
\title{Evaluates the negeative log-likelihood for model BinaryT1:S with lambda depending 
on a covariate.}
\usage{
nll.binT1S(param, R, Tmax, dat)
}
\arguments{
\item{param}{A vector comprised of parameters \eqn{b_0}{b0}, \eqn{b_1}{b1}, 
and the log of the detection hazard \eqn{h}{h} (in that order), where the log of the Poisson rate 
lambda is equal to \eqn{b_0+b_1x}{b0+b1*x} and \eqn{x}{x} is the coaviate \code{covar}.}

\item{R}{The number of sites.}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}

\item{dat}{An \code{R} by 1 matrix of binary values, with a 1 representing
detection and a 0 representing no detection}

\item{covar}{A vector covariate of length \code{R} on which the expected number of 
animals in the site depends linearly (assumed to be the same for all occasions).}
}
\value{
Returns the negative log-likelihood function evaluated for model BinaryT1:S at 
the parameter values passed in \code{param}.
}
\description{
Evaluates the negeative log-likelihood for model BinaryT1:S, assuming that lambda 
is constant, given initial parameter estimates and binary data from a 
single-occasion survey.
}
\examples{
Rsites=100. #number of sites
Tsearch=3 #maximum time

#true parameters
#true parameters
lamt = 2
ht=0.4620981
paramt=c(lamt,ht)

# data for BinaryT1:S
bns=as.matrix(generate.binT1S(paramt,R=Rsites,Tmax=Tsearch))

init.paramt=c(log(lamt),log(ht))
nll = nll.binT1S(param=init.paramt, R=Rsites, Tmax=Tsearch,dat=bns)
nll

# optimize
set.seed(123) # for reproducibility
bn1s=as.matrix(generate.binT1S(paramt,R=Rsites,Tmax=Tsearch))
# optimize
fit.bin1S=optim(init.paramt,nll.binT1S,R=Rsites,Tmax=Tsearch,dat=bn1s)
estpar.bin1S=fit.bin1S$par
# compare estimates and true parameters
exp(estpar.bin1S)
paramt 

}
