% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countM.R
\name{nll.countM}
\alias{nll.countM}
\title{Evaluates the negeative log-likelihood for model Count:M with lambda constant}
\usage{
nll.countM(param, R, J, Tmax, dat)
}
\arguments{
\item{param}{A vector comprised of the log of the Poisson rate lambda, and the 
log of the detection hazard, h.}

\item{R}{The number of sites.}

\item{J}{The number of occasions (assumed the same for all sites).}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}

\item{dat}{An \code{R} by \code{J} matrix of counts.}
}
\value{
Returns the negative log-likelihood function evaluated at the parameter values
passed in \code{param}.
}
\description{
Evaluates the negeative log-likelihood for model Count:M, assuming constant lambda, 
given initial parameter estimates and count data from a multiple-occasion survey.
}
\examples{
# setting
Rsites=100  #number of sites
Jsites=5    #multiple visits
Tsearch=3 #maximum time
# true parameters
lamt = 2
ht=0.4620981
paramt=c(lamt,ht)

# data for Count:M
set.seed(123) # for reproducibility
cntm=as.matrix(generate.countM(paramt,R=Rsites,J=Jsites,Tmax=Tsearch))
# optimize
init.paramt=c(log(lamt),log(ht))
fit.countM=optim(init.paramt,nll.countM,R=Rsites,J=Jsites,Tmax=Tsearch,dat=cntm)
estpar.countM=fit.countM$par
# compare estimates and true parameters
exp(estpar.countM)
paramt 

}
