% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMT1.R
\name{nll.countMT1}
\alias{nll.countMT1}
\title{Evaluates the negeative log-likelihood for model CountT:M with lambda constant}
\usage{
nll.countMT1(param, R, J, Tmax, dat)
}
\arguments{
\item{param}{A vector comprised of the log of the Poisson rate lambda, and the 
log of the detection hazard, h.}

\item{R}{The number of sites.}

\item{J}{The number of occasions (assumed the same for all sites).}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}

\item{dat}{An \code{R} by \code{2J} matrix with first \code{J} columns being the numbers 
of detections in each site on each occasion and the remaining \code{J} columns being the 
times to first detection of an animal within the site on the occasion 
(with zero representing no detections).}
}
\value{
Returns the negative log-likelihood function evaluated at the parameter values
passed in \code{param}.
}
\description{
Evaluates the negeative log-likelihood for model CountT:M, assuming constant lambda, 
given initial parameter estimates and count data from a multiple-occasion survey.
}
\examples{
# setting
Rsites=100  #number of sites
Jsites=5    #multiple visits
Tsearch=3 #maximum time
# true parameters
lamt = 2
ht=0.4620981
paramt=c(lamt,ht)

# data for CountT1:M
set.seed(123) # for reproducibility
cntm=as.matrix(generate.countMT1(paramt,R=Rsites,J=Jsites,Tmax=Tsearch))
# optimize
init.paramt=c(log(lamt),log(ht))
fit.countMT1=optim(init.paramt,nll.countMT1,R=Rsites,J=Jsites,Tmax=Tsearch,dat=cntm)
estpar.countMT1=fit.countMT1$par
# compare estimates and true parameters
exp(estpar.countMT1)
paramt 

}
