% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PcountM.R
\name{nll.PcountM}
\alias{nll.PcountM}
\title{Evaluates the negeative log-likelihood for model PCount:M with lambda constant.}
\usage{
nll.PcountM(param, R, J, Tmax, dat)
}
\arguments{
\item{param}{A vector comprised of the Poisson rate lambda, and the detection hazard, h.}

\item{R}{The number of sites.}

\item{J}{The number of occasions.}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}

\item{dat}{An \code{R} by \code{J} matrix of counts}
}
\value{
Returns the negative log-likelihood function evaluated for model PCount:M at 
the parameter values passed in \code{param}.
}
\description{
Evaluates the negeative log-likelihood for model PCount:M, assuming that lambda 
is constant, given initial parameter estimates and binary data from a multiple-occasion 
survey.
}
\examples{
# seed
set.seed(123)
# setting
Rsites=100. #number of sites
Tsearch=3 #maximum time
Joccs=5 #number of occasions
#true parameters
lamt=2
gt=0.4620981
paramt=c(lamt,gt)

ymat=generate.PcountM(paramt, R=Rsites, J=Joccs, Tmax=Tsearch)
str(ymat)
head(ymat)

param0=c(log(lamt),log(gt))
# evaluate negative log-likelihood at starting values:
nll.PcountM(param0, R=Rsites, J=Joccs, Tmax=Tsearch, dat=ymat)

# optimize
fit.PcountM=optim(param0, nll.PcountS, R=Rsites, J=Joccs, Tmax=Tsearch, 
                  dat=ymat, method="BFGS")
estpar.PcountM=fit.PcountM$par
# compare estimates and true parameters
exp(estpar.PcountM)
paramt 

}
