% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PcountS.R
\name{nll.PcountS}
\alias{nll.PcountS}
\title{Evaluates the negeative log-likelihood for model PCount:S with lambda constant.}
\usage{
nll.PcountS(param, R, Tmax, dat)
}
\arguments{
\item{param}{A vector comprised of the Poisson rate lambda, and the detection hazard, h.}

\item{R}{The number of sites.}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}

\item{dat}{An \code{R} by 1 matrix of counts}
}
\value{
Returns the negative log-likelihood function evaluated for model PCount:S at 
the parameter values passed in \code{param}.
}
\description{
Evaluates the negeative log-likelihood for model PCount:S, assuming that lambda 
is constant, given initial parameter estimates and binary data from a multiple-occasion 
survey.
}
\examples{
# seed
set.seed(123)
# setting
Rsites=100. #number of sites
Tsearch=3 #maximum time
#true parameters
lamt=2
gt=0.4620981
paramt=c(lamt,gt)

ymat=generate.PcountS(paramt, R, Tmax)
str(ymat)
head(ymat)

param0=c(log(lamt),log(gt))
# evaluate negative log-likelihood at starting values:
nll.PcountS(param0, R=Rsites, Tmax=Tsearch, dat=ymat)

# optimize
fit.PcountS=optim(param0, nll.PcountS, R=Rsites, Tmax=Tsearch, dat=ymat, method="BFGS")
estpar.PcountS=fit.PcountS$par
# compare estimates and true parameters
exp(estpar.PcountS)
paramt 

}
