% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmtd.R
\docType{package}
\name{nmitd}
\alias{nmtd}
\alias{nmtd-package}
\title{nmtd: A package for N-mixture models with Time-to-Detection Data}
\description{
Functions for data generation and fitting of N-mixture models with Time-to-Detection Data
}
\references{
Altwegg, R., Borchers, D.L. and Haines, L.M. 2023 Exact Likelihoods for N-mixture models 
with Time-to-Detection Data. ANZJS.
}
\author{
Linda Haines.
}
\keyword{N-mixture,}
\keyword{Time}
\keyword{detection}
\keyword{to}
