% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Karoodata}
\alias{Karoodata}
\title{Data from survey in the Karoo}
\format{
A data frame with 4015 observations on the following 19 variables.
 \describe{
   \item{\code{survey_id}}{tbc}
   \item{\code{temperature}}{tbc}
   \item{\code{cloud}}{tbc}
   \item{\code{wind}}{tbc}
   \item{\code{rain}}{tbc}
   \item{\code{pentad}}{tbc}
   \item{\code{survey_start}}{tbc}
   \item{\code{datetime}}{tbc}
   \item{\code{species}}{tbc}
   \item{\code{ttd}}{tbc}
   \item{\code{longitude}}{tbc}
   \item{\code{latitude}}{tbc}
   \item{\code{mins_after_dawn seenheard}}{tbc}
   \item{\code{perchedflying}}{tbc}
   \item{\code{distance}}{tbc}
   \item{\code{latin_name}}{tbc}
   \item{\code{date}}{tbc}
   \item{\code{time}}{tbc}
 }
}
\source{
https://doi.org/10.5061/dryad.msbcc2fv7.
}
\usage{
Karoodata
}
\description{
Data from survey in the Karoo.
}
\details{
To be completed.
}
\examples{
 data(Karoodata)
}
\references{
Henry, Dominic; Lee, Alan; Altwegg, Res (2020), Can time-to-detection models with fewer 
survey replicates provide a robust alternative to traditional site-occupancy models? 
Methods in Ecology and Evolution, 11: 643-655. 
https://besjournals.onlinelibrary.wiley.com/doi/10.1111/2041-210X.13379
}
