% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binM.R
\name{nll.binM}
\alias{nll.binM}
\title{Evaluates the negeative log-likelihood for model Binary:M with lambda constant}
\usage{
nll.binM(param, R, J, Tmax, dat)
}
\arguments{
\item{param}{A vector comprised of the log of the Poisson rate lambda, and the 
log of the detection hazard, h.}

\item{R}{The number of sites.}

\item{J}{The number of occasions (assumed the same for all sites).}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}

\item{dat}{An \code{R} by \code{J} matrix of binary values, with a 1 representing
detection and a 0 representing no detection}
}
\value{
Returns the negative log-likelihood function evaluated at the parameter values
passed in \code{param}.
}
\description{
Generates Poisson random variables for each of the \code{R} sites, corresponding to the 
number of animals in each site. Then generates Bernoulli random variables for each of 
\code{J} occasions that each site is surveyed, assuming a constant hazard of detection 
for each animal.
}
\examples{
# setting
Rsites=100  #number of sites
Jsites=5    #multiple visits
Tsearch=3 #maximum time
#true parameters
lamt = 2
ht=0.4620981
paramt=c(lamt,ht)
param0=c(log(lamt),log(ht))

bnm=as.matrix(generate.binM(paramt,R=Rsites,J=Jsites,Tmax=Tsearch))
# optimize
finopt=optim(param0,nll.binM,R=Rsites,J=Jsites,Tmax=Tsearch,dat=bnm)
finpar=finopt$par
# save
exp(finpar[1])
exp(finpar[2])

}
