% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binS.R
\name{nll.binS}
\alias{nll.binS}
\title{Evaluates the negeative log-likelihood for model Binary:S with lambda depending 
on a scalar covariate.}
\usage{
nll.binS(param, R, J, Tmax, dat, covar)
}
\arguments{
\item{param}{A vector comprised of parameters \eqn{b_0}{b0}, \eqn{b_1}{b1}, and the log
of the detection hazard (in that order), where the log of the Poisson rate lambda is equal to 
\eqn{b_0+b_1x}{b0+b1*x} and \eqn{x}{x} is the coaviate \code{covar}.}

\item{R}{The number of sites.}

\item{J}{The number of occasions (assumed the same for all sites).}

\item{Tmax}{The survey duration (assumed to be the same for all sites)}

\item{dat}{An \code{R} by \code{J} matrix of binary values, with a 1 representing
detection and a 0 representing no detection}

\item{covar}{A (scalar) covariate on which the hazard depends linearly (assumed to be 
the same for all sites).}
}
\value{
Returns the negative log-likelihood function evaluated at the parameter values
passed in \code{param}.
}
\description{
Generates Poisson random variables for each of the \code{R} sites, corresponding to the 
number of animals in each site. Then generates Bernoulli random variables for each of 
\code{J} occasions that each site is surveyed, assuming a constant hazard of detection 
for each animal.
}
\examples{
###### NOT YET UPDATED #########
# setting
Rsites=100  #number of sites
Jsites=5    #multiple visits
Tsearch=3 #maximum time
#true parameters
lamt = 2
ht=0.4620981
paramt=c(lamt,ht)
param0=c(log(lamt),log(ht))

bnm=as.matrix(generate.binM(paramt,R=Rsites,J=Jsites,Tmax=Tsearch))
# optimize
finopt=optim(param0,nll.binM,R=Rsites,J=Jsites,Tmax=Tsearch,dat=bnm)
finpar=finopt$par
# save
exp(finpar[1])
exp(finpar[2])

}
